refact(tui): dedupe Ctrl+T overlay and move transcript builder
