+++ codex-rs/tui/src/app_backtrack.rs
@@ -9,6 +9,7 @@ use color_eyre::eyre::Result;
 use crossterm::event::KeyCode;
 use crossterm::event::KeyEvent;
 use crossterm::event::KeyEventKind;
+use ratatui::text::Line;
 
 /// Aggregates all backtrack-related state used by the App.
 #[derive(Default)]
@@ -341,5 +342,29 @@ impl App {
         } else if !lines.is_empty() {
             // If no cut index found (e.g., drop_count == 0), render nothing.
         }
+    }
+
+    /// Build flattened transcript lines and absolute user spans on demand.
+    /// This replaces the previous persistent `transcript_lines`/`user_spans` state.
+    pub(crate) fn build_transcript_flattened(&self) -> (Vec<Line<'static>>, Vec<(usize, usize)>) {
+        let mut out: Vec<Line<'static>> = Vec::new();
+        let mut spans: Vec<(usize, usize)> = Vec::new();
+        for cell in &self.transcript_cells {
+            let is_stream = cell.is_stream_continuation();
+            let mut lines = cell.transcript_lines();
+            if !is_stream && !out.is_empty() && !lines.is_empty() {
+                out.push("".into());
+            }
+            let start = out.len();
+            if let Some(span) = cell.message_span()
+                && matches!(cell.kind(), crate::history_cell::MessageKind::User)
+            {
+                let header_abs = start.saturating_add(span.header_offset);
+                let end_abs = header_abs.saturating_add(1).saturating_add(span.body_len);
+                spans.push((header_abs, end_abs));
+            }
+            out.append(&mut lines);
+        }
+        (out, spans)
     }
 }
